+++ src/unittest/sort.cpp
@@ -178,7 +178,7 @@ TEST_CASE("Sorting a graph with paths 1 node long", "[sort]") {
     std::function<uint64_t(const std::vector<handlegraph::path_handle_t> &,
                            const xp::XP &)> get_max_path_step_count
             = [&](const std::vector<handlegraph::path_handle_t> &path_sgd_use_paths, const xp::XP &path_index) {
-                uint64_t max_path_step_count = 0;
+                size_t max_path_step_count = 0;
                 for (auto& path : path_sgd_use_paths) {
                     max_path_step_count = std::max(max_path_step_count, path_index.get_path_step_count(path));
                 }
